Added position last
git-svn-id: https://webgate.ec.europa.eu/CITnet/svn/UNIONVMS/branches/dev/Modules/Reporting/APP@5492 ffa9caa6-9a2d-11e4-8c71-fd203efb3018

+++ service/src/main/java/eu/europa/ec/fisheries/uvms/reporting/service/dto/PositionSelectorDTO.java
@@ -10,24 +10,24 @@ public class PositionSelectorDTO {
 
     public final static String VALUE = "value";
 
-    private Long value;
+    private Float value;
 
     private Selector selector;
 
     private Position position;
 
     @Builder(builderMethodName = "PositionSelectorDTOBuilder")
-    public PositionSelectorDTO(Long value, Selector selector, Position position) {
+    public PositionSelectorDTO(Float value, Selector selector, Position position) {
         this.value = value;
         this.selector = selector;
         this.position = position;
     }
 
-    public Long getValue() {
+    public Float getValue() {
         return value;
     }
 
-    public void setValue(Long value) {
+    public void setValue(Float value) {
         this.value = value;
     }
